requirements.txt
numpy 
opencv-python
argparse
pillow
